Remove old ESI syncKillboard
Moves syncZKillboard contents into syncKillboard, fixes cron reference,
removes debug log statement that was missed, and sets user-agent based
on config file.
