+++ back-end/src/cron/cron.js
@@ -20,7 +20,7 @@ const TASKS = [
   },
   {
     name: 'syncKillboard',
-    executor: require('./task/syncZKillboard'),
+    executor: require('./task/syncKillboard'),
     timeout: moment.duration(30, 'minutes').asMilliseconds(),
     interval: moment.duration(1, 'day').asMilliseconds(),
     schedule: '0 2 * * *',  // Once a day at 2AM
